refactor: split different sources
