+++ .pre-commit-config.yaml
@@ -22,34 +22,23 @@ repos:
             entry: black
             language: system
             types: [python]
-            exclude: |
-                (?x)^(
-                    docs/*
-                )$
+            exclude: ^docs/
+            verbose: true
           - id: yapf
             name: yapf
             entry: yapf -r -i
             language: system
             types: [python]
-            exclude: |
-                (?x)^(
-                    docs/*
-                )$
+            exclude: ^docs/
           - id: pylint
             name: pylint
             entry: pylint
             language: system
             types: [python]
-            exclude: |
-                (?x)^(
-                    docs/*
-                )$
+            exclude: ^docs/
           - id: mdformat
             name: mdformat
             entry: mdformat
             language: system
             types: [markdown]
-            exclude: |
-                (?x)^(
-                    CHANGELOG.md
-                )$
+            exclude: ^CHANGELOG.md
