+++ graph_datasets/datasets/__init__.py
@@ -1,0 +1,8 @@
+"""Datasets from different sources
+"""
+from .cola import load_cola_data
+from .dgl import load_dgl_data
+from .linkx import load_linkx_data
+from .ogb import load_ogb_data
+from .pyg import load_pyg_data
+from .sdcn import load_sdcn_data
